Generate corpus/models.py
Models for the corpus app have been generated based on the existing
database supplied by Nederlab. This database contains the information
that needs to be visualized. Because we do not want to edit this
information, the generated models.py is used as is (no updates for
foreign keys, etc.). Everything in the database seems to be saved in
text-fields.