Add author details to list of plays
The author details (name, etc.) are displayed in the list of plays.
These details are stored in table Auteur. Authors and books are
linked via table Titelxauteur. Author details are retrieved
manually, because it is not clear how Django makes these connections
between tables. This is relatively inefficient, but it is probably more
work to fix it than to work around it.