Fit nederlab-dbnl database in Django format (titel and auteur)
Using the Django database management system with an existing
database proved less of a fuss than expected. Because this makes the
code cleaner and more efficient, the many to many mapping between titel
and auteur was changed to use Django database management. The code for
displaying the list of plays was also updated.

Next, the other relevant many to many relations need to be changed as
well. Including genre and subgenre.