Add subgenre histogram to genres page
A histogram for subgenres is displayed on the genres page. Subgenres
that occur at least 5 times are included in the histogram.

To properly display the subgenres, 8 distinct colors are needed (there
are 8 subgenres that occur at least 5 times). At the moment there are
only 7 colors available. Also the order of the genres and subgenres in
the bars needs to be changed from seemingly random to based on
frequency.