Add multiple LIWC categories to graph
Based on a new data file that specifies the frequencies of multiple
LIWC categories (of a single speaker), a new line graph is created.
The graph shows frequencies per sentence for all LIWC categories in
the data.
Next step: smoothing (the lines currently show raw frequencies and need
to be smoothed).