Update relation between SpeakingTurn and Entity
The relation between SpeakingTurn and Entity was a ForeingKey
relationship. This is a conceptual mistake: a SpeakingTurn can have
many Entities and many Entities can occur in a SpeakingTurn. Therefore,
the relation between SpeakingTurn and Entity was changed to a
ManyToMany relation. Because the entity score must also be saved, a new
model EntityScore was added to the entity_vis models for this relation.