Add management command to store EntityScores in the database
Added a management command that extracts Characters, Entities,
SpeakingTurns and EntityScores from a FoLiA XML file and add them to
the database. The information in the database is used for
visualization.
What needs to be improved: at the moment it is still possible that the
same speaking turn/entity score is added multiple times. This needs to
be avoided. Also, it might be a good idea to add a field 'class' to the
Entity model (now an Entity.name consists of <class>-<name>.