Retrieve data for visualization from database
The entities_in_play view function now retrieves the data that should
be visualized from the database, puts it in the format expected by the
JavaScript script and returns it to the template.
Next, the JavaScript script should be adapted to use the data returned
by the view instead of data from a csv file.