Add EntityValue model
A model for storing the words associated with entities was added
(EntityValue) to the entity_vis app. The words associated with entities
will be displayed together with the graph visualizing the entities.
The words will be displayed per speaking turn. Therefore, a
ManyToManyField to EntityValue was added to the EntityScore model.