Fix corpus/models
The foreign key relation between the entity_vis Character model and the
corpus Titel model caused a MySQL problem, because the existing DBNL
data (of the corpus app) was stored with the MyISAM storage engine,
while Django's default is INNODB. This caused errors when generating the
table for the Character model. To solve this problem, we made Django
generate the database tables for the corpus app. In order to do that
some changes to corpus/models.py were made. This commit records these
changes.
To solve the MyISAM/INNODB problem, the tables were generated by Django
and afterwards that the data was loaded into the tables.