Add foreign keys to TitelBevat model
The TitelBevat model contains a mapping from titles to the titles it
contains. Using this mapping, we can find the first year of publication
of a title, instead of some later year (when the title was republished
later). We want to use this first year of publication in visualizations,
etc. In order to be able to use the TitelBevat table/model for these
purposes, the foreign key relations to the Titel table/model were
updated. This step also involved reloading of the data in the table,
because there were links to non-existing titles. These have been
removed.