Update Titel model to use TitelBevat directly
The Titel model was updated to have TitelBevat as a many to many
relation. This has the benefit that titles contained in a title can be
found directly through the current title, instead of though a TitelBevat
object. The corresponding view and template were updated to use this new
direct relationship.