Add management command to save the year to events in ES
A management command that adds a year field to the events in
ElasticSearch was added to the corpus app. The command finds all
text_ids in the ElasticSearch index and looks up the first year of
publication of a title. Than all events are retrieved for a title and
a new field containing the year as an integer is added. If a title
contains multiple titles, no years are added, because it is not clear
which title is needed. This somehow needs to be fixed later. In the
collection of titles selected for annotation no multiple titles are
contained, so the problem does not occur.