Add url mapping and view for displaying information about speakingturns
Added an url mapping and view for displaying information about
speakingturns. Speakingturns are selected based on a list of words. The
list of words is still hardcoded, this needs to be changed later. The
information extracted is the number of documents (events) for each word
in the list, the speakingturns containing the word, and word frequencies
of words in the speakingturns.