Update url mapping and view to display info for different concepts
Previously, the find_in_speakingturns view returned information about a
list of words that was hardcoded. The view now contains a dictionary
containing lists of words that are spelling variants of a concept. The
url mapping was updated to accept an additional parameter that specifies
the concept to be searched for. Based on this argument, the view
returns information for the appropriate list of spelling variants.