Add management command to print a list of available texts
Added a management command that prints a list of ids of titles whose
text is available in Nederlab/DBNL. The command takes an input parameter
specifying a list of titles in Nederlab/DBNL. The command filters this
list by checking whether a title is a play (genre drama) and the year is
in the right time period.
The next step is to also include titles via title.contains, to find
newer publications of old plays.