Add all titles using title.contains
The find_texts command now also checks whether titles contained in the
list of available titles are of genre drama and come from the right time
period. If so, the available title is added to the list of titles to
request. Sometimes a title has multiple versions in the relevant time
period, so doubles are added. They must be removed later.