Update entities_in_play view to retrieve data from elasticsearch
The entities_in_play view now retrieves its data directly from
elasticsearch. Currently, the raw frequencies of Posemo tags are
retrieved for the top 5 characters of alew001besl01. Later, the
hardcoded parameters must be made variable, the results must be smoothed
and a measure must be calculated for the results.