Add url mapping and view to retrieve subgenre data over time
Added an url mapping and a view to retrieve data on the percentages of
words that belong to some entity category for different subgenres over
time. The entity category is hardcoded to 'Posemo'. This should be made
dependent on the selected categories (a graph should be generated per
category). Also, a graph displaying the number of texts in the corpus
should be placed on top.