Refactor data returned by the subgenres_stats_time view
The data returned by the subgenres_stats_time view used custom names for
the data points (year and percentage). This was changed to the more
generic names x and y, so the chart function can use a generic function
to find x and y coordinates. This prevents repeating functions to
extract the x and y values from the data.