Fix year2es and genre2es for titles with multiple subtitles
The year and genre data of titles with multiple subtitles were not added
to the ES index, because it is unclear what data to use. Now the details
from the original title are used instead.
Also added a utility function that extracts the year from a string using
a regular expression (to deal with years like 'ca. 1768').