Fix removing data from graph
The final series was not removed from the graph, although the graph
displayed the message 'no data available' after the last pair was
removed. This issue was fixed by adding a unique id to the chart.

A unique id should be added to all charts.

Also when adding new pairs to an emptied chart, the legend colors don't
match anymore. This should also be fixed.