Add scatter plot for corpus
Added a scatter plot that displays the percentages of posemo tags for
each text in the corpus. Next steps are to make the entity categories
displayed dynamic (currently it is hardcoded posemo) and to add scatter
plots for the genre data.
Also, the scatter data must be loaded automatically (currently, a button
has to be clicked)